Make opls/trappe global fixtures, Add tests for isomorphism
