+++ gmso/tests/parameterization/test_trappe_gmso.py
@@ -65,14 +65,6 @@ def _match_connection_parameters(
 
 class TestTrappeGMSO(BaseTest):
     @pytest.fixture(scope="session")
-    def xml_loader(self):
-        return FoyerFFs()
-
-    @pytest.fixture(scope="session")
-    def trappe_ua_gmso(self, xml_loader):
-        return xml_loader.load("trappe_ua", rel_to_module=True).to_gmso_ff()
-
-    @pytest.fixture(scope="session")
     def trappe_ua_openmm_foyer(self):
         return foyer.forcefields.load_TRAPPE_UA()
